+++ tests/Pest.php
@@ -2,7 +2,6 @@
 declare( strict_types = 1 );
 
 define('MARMERINE_PORT', (int) getenv('MARMERINE_PORT') ?: 11211);
-
 echo 'PORT: ' . MARMERINE_PORT . PHP_EOL;
 
 class MC {